backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
mediapipe==0.10.14
numpy==1.26.4
pillow==10.4.0
torch==2.4.1
torchvision==0.19.1
transformers==4.48.0
huggingface_hub==0.28.0
